Add support for simple_build.txt options file
If a file "simple_build.txt" exists in the
root directory, simple_build.sh will consider
if an "options" file. It takes arguments, parses
them and feeds the content to the build system
arguments. This file also supports comments.
By default, comments start with '#'. They are
parsed with sed, so special sed keywords should
not be used.