Rewrite i3_resurrect
i3_resurrect_script.sh is now a wrapper for the i3-resurrect command. The
reason behind this decision is the command lacks some features, like listing
all profiles and saving/restoring/removing all workspaces. The i3-resurrect
project uses the Python Click library, which is not straight forward to work
with. Creating a wrapper for the command instead of forking it is therefore
a lot easier to work with in Bash.

Commands added to the script:
- Save all workspaces. Only used workspaces are saved
- Restore all workspaces. For now, this also removes them as they are restored
- List all saved workspaces or all profiles